feat: useRegister로 terms state 분리 및 이용약관 페이지 출력 방식 변경
